+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## Version 1.6 (development)
 
+* Fix compatibility with Vera's Amazon Alexa integration (which appends temp unit on action parmeter).
 * Fix so odd return values from underlying Luup requests don't crash Status Data request.
 * Experimental fix for openLuup switch control (stable-180918)
 * Upgrade detection of AltUI so we don't falsely detect when bridged (on "real" device triggers AltUI feature registration).
-* Implement AllSetpoints state variable, for user convenience. 
+* Implement AllSetpoints state variable, for user convenience (and Alexa)
 * openLuup loads modules differently from Vera Luup, so work out how to get AVT working on openLuup when multiple device instances are created (differences don't matter if only one device).
 
 ## Version 1.5 (released) ##